fix(style): fix some styles and refactor some files
