+++ src/GuppyDataExplorer/ExplorerButtonGroup/ExplorerButtonGroup.css
@@ -1,0 +1,3 @@
+.explorer-button-group__download-button {
+  margin-right: 10px;
+}
