+++ src/GuppyDataExplorer/ExplorerTable/ExplorerTable.css
@@ -1,0 +1,7 @@
+.explorer-table {
+  margin-top: 20px;
+}
+
+.explorer-table__description{
+  margin-bottom: 10px;
+}
