+++ src/GuppyDataExplorer/ExplorerTable/index.jsx
@@ -2,6 +2,8 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import ReactTable from 'react-table';
 import 'react-table/react-table.css';
+import { GuppyConfigType } from '../configTypeDef';
+import './ExplorerTable.css';
 
 class ExplorerTable extends React.Component {
   constructor(props) {
@@ -9,6 +11,7 @@ class ExplorerTable extends React.Component {
     this.state = {
       loading: false,
       pageSize: props.defaultPageSize,
+      currentPage: 0,
     };
   }
 
@@ -27,29 +30,33 @@ class ExplorerTable extends React.Component {
       this.setState({
         loading: false,
         pageSize: size,
+        currentPage: state.page,
       });
     });
   }
 
   render() {
-    console.log(this.props.rawData);
     const columnsConfig = this.props.tableConfig.map(c => ({ Header: c.name, accessor: c.field }));
     const { totalCount } = this.props;
     const { pageSize } = this.state;
     const totalPages = Math.floor(totalCount / pageSize) + ((totalCount % pageSize === 0) ? 0 : 1);
+    const start = this.state.currentPage * this.state.pageSize + 1;
+    const end = (this.state.currentPage + 1) * this.state.pageSize;
     return (
-
-      <ReactTable
-        columns={columnsConfig}
-        manual // Forces table not to paginate or sort automatically, so we can handle it server-side
-        data={this.props.rawData}
-        pages={totalPages} // Display the total number of pages
-        loading={this.state.loading} // Display the loading overlay when we need it
-        onFetchData={this.fetchData.bind(this)} // Request new data when things change
-        defaultPageSize={this.props.defaultPageSize}
-        // className={`-striped -highlight `}
-        minRows={0} // hide empty rows
-      />
+      <div className='explorer-table'>
+        <p className='explorer-table__description'>{`Showing ${start} - ${end} of ${totalCount} cases`}</p>
+        <ReactTable
+          columns={columnsConfig}
+          manual // Forces table not to paginate or sort automatically, so we can handle it server-side
+          data={this.props.rawData}
+          pages={totalPages} // Display the total number of pages
+          loading={this.state.loading} // Display the loading overlay when we need it
+          onFetchData={this.fetchData.bind(this)} // Request new data when things change
+          defaultPageSize={this.props.defaultPageSize}
+          // className={`-striped -highlight `}
+          minRows={0} // hide empty rows
+        />
+      </div>
     );
   }
 }
@@ -64,6 +71,7 @@ ExplorerTable.propTypes = {
     field: PropTypes.string,
     name: PropTypes.string,
   })).isRequired,
+  guppyConfig: GuppyConfigType,
 };
 
 ExplorerTable.defaultProps = {
